renamed stake sale to user_unstaked
