+++ v0_4/src/lib.rs
@@ -21,7 +21,7 @@ pub mod node_config;
 pub mod rewards;
 pub mod node_activation;
 pub mod pause;
-pub mod stake_sale;
+pub mod user_unstake;
 pub mod unexpected;
 pub mod user_data;
 pub mod user_stake;
@@ -36,7 +36,7 @@ use crate::rewards::*;
 use crate::node_activation::*;
 use crate::pause::*;
 use crate::user_stake::*;
-use crate::stake_sale::*;
+use crate::user_unstake::*;
 use crate::unexpected::*;
 use crate::user_data::*;
 use crate::fund_transf_module::*;
@@ -79,7 +79,7 @@ pub trait Delegation {
     fn node_activation(&self) -> NodeActivationModuleImpl<T, BigInt, BigUint>;
 
     #[module(StakeSaleModuleImpl)]
-    fn stake_sale(&self) -> StakeSaleModuleImpl<T, BigInt, BigUint>;
+    fn user_unstake(&self) -> StakeSaleModuleImpl<T, BigInt, BigUint>;
 
     #[module(UnexpectedBalanceModuleImpl)]
     fn unexpected(&self) -> UnexpectedBalanceModuleImpl<T, BigInt, BigUint>;
